Create Fortune 500 Analysis
This was an open-ended project intended to give me the chance to think of different ways I could select the data, just like you would on the job as a data analyst.

For this project, I used a database that has fake Fortune 500 data in it. The Fortune 500 is a list of the 500 companies with the most revenue and that are typically regarded as some of the best places to work.

My task was to write at least 5 different queries.
Do each of the following at least once:

Group the data in a new way using a CASE statement.
Use a HAVING clause to determine something interesting about the data per category.
Use logical operators like AND or OR to filter the data in an interesting way.
Use an aggregate function like AVG, SUM, COUNT, MAX, and/or MIN to return summary statistics about the data.